requirements.txt
anyio==3.7.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
httptools==0.5.0
idna==3.4
instaloader==4.9.6
outcome==1.2.0
packaging==23.1
pydantic==1.10.9
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
webdriver-manager==3.8.6
websockets==11.0.3
wsproto==1.2.0